+++ packages/create-orderly-app/CHANGELOG.md
@@ -1,5 +1,31 @@
 # create-orderly-app
 
+## 1.7.0
+
+### Minor Changes
+
+- e1badae: publish
+- publish
+
+## 1.7.0-alpha.0
+
+### Minor Changes
+
+- publish
+
+## 1.6.3
+
+### Patch Changes
+
+- publish
+- 5ba2a31a8: publish
+
+## 1.6.3-alpha.0
+
+### Patch Changes
+
+- publish
+
 ## 1.6.2
 
 ### Patch Changes