+++ packages/default-solana-adapter/CHANGELOG.md
@@ -1,5 +1,51 @@
 # @orderly.network/default-solana-adapter
 
+## 2.2.0
+
+### Minor Changes
+
+- e1badae: publish
+- publish
+
+### Patch Changes
+
+- Updated dependencies [e1badae]
+- Updated dependencies
+  - @orderly.network/core@2.2.0
+  - @orderly.network/types@2.2.0
+
+## 2.2.0-alpha.0
+
+### Minor Changes
+
+- publish
+
+### Patch Changes
+
+- Updated dependencies
+  - @orderly.network/core@2.2.0-alpha.0
+  - @orderly.network/types@2.2.0-alpha.0
+
+## 2.1.3
+
+### Patch Changes
+
+- publish
+- 5ba2a31a8: publish
+- Updated dependencies
+- Updated dependencies [5ba2a31a8]
+  - @orderly.network/core@2.1.3
+  - @orderly.network/types@2.1.3
+
+## 2.1.3-alpha.0
+
+### Patch Changes
+
+- publish
+- Updated dependencies
+  - @orderly.network/core@2.1.3-alpha.0
+  - @orderly.network/types@2.1.3-alpha.0
+
 ## 2.1.2
 
 ### Patch Changes