+++ packages/net/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @orderly.network/net
 
+## 2.2.0
+
+### Minor Changes
+
+- e1badae: publish
+- publish
+
+## 2.2.0-alpha.0
+
+### Minor Changes
+
+- publish
+
+## 2.1.3
+
+### Patch Changes
+
+- publish
+- 5ba2a31a8: publish
+
+## 2.1.3-alpha.0
+
+### Patch Changes
+
+- publish
+
 ## 2.1.2
 
 ### Patch Changes