+++ packages/tsconfig/CHANGELOG.md
@@ -1,5 +1,31 @@
 # tsconfig
 
+## 0.5.0
+
+### Minor Changes
+
+- e1badae: publish
+- publish
+
+## 0.5.0-alpha.0
+
+### Minor Changes
+
+- publish
+
+## 0.4.3
+
+### Patch Changes
+
+- publish
+- 5ba2a31a8: publish
+
+## 0.4.3-alpha.0
+
+### Patch Changes
+
+- publish
+
 ## 0.4.2
 
 ### Patch Changes